requirements.txt
aiofiles==23.2.1
aiohttp==3.9.2
apscheduler==3.6.3
asyncio==3.4.3
beautifulsoup4==4.12.3
bing_image_downloader==1.1.2
bleach==6.1.0
bs4==0.0.2
cachetools==4.2.2
cloudscraper==1.2.71
countryinfo==0.1.2
currencyconverter==0.17.20
dateparser==1.2.0
emoji==2.11.1
emoji-country-flag==1.3.2
feedparser==6.0.11
ffmpeg-python==0.2.0
future==0.18.3
geopy==2.4.1
gpytranslate==1.5.1
hachoir==3.3.0
html2text==2024.2.26
jikanpy==4.3.2
lxml==4.9.4
markdown2==2.4.13
motor==3.3.2
opencv-python-headless==4.9.0.80
pillow==9.5.0
pretty_errors==1.2.25
psutil==5.9.8
psycopg2-binary==2.9.9
pydictionary==2.0.1
pykeyboard==0.1.5
pynewtonmath==1.0.1
pyrate-limiter==2.10.0
pyrogram==2.0.106
python-telegram-bot==13.15
regex==2023.12.25
requests==2.31.0
secureme==0.0.4
search_engine_parser==0.6.8
speedtest-cli==2.1.3
sqlalchemy==1.4.46
telegraph==2.2.0
telethon==1.34.0
tgcrypto==1.2.5
wikipedia==1.4.0

# update modules extra

gtts
heroku3
alphabet_detector
wget
nekos.py
yt_dlp
youtube_search
youtube_search_python
pyshorteners
faker
pyfiglet
MukeshAPI
unidecode
openai==1.31.0
numpy
pycountry
blackpink 
qrcode
lexica-api==1.5.7
git+https://github.com/AsmSafone/SafoneAPI
pydub
pymongo==4.6.3
httpx==0.27.0
img2pdf
dnspython==2.2.1
pyromod==1.5
pyrogram1==0.1.0
python-dotenv==0.21.0
